make duplicate embedding_query_content function to util and add truncate by token
