+++ autorag/nodes/retrieval/vectordb.py
@@ -58,6 +58,12 @@ async def vectordb_pure(queries: List[str], top_k: int, collection: chromadb.Col
 
     :return: The tuple contains a list of passage ids that retrieved from vectordb and a list of its scores.
     """
+    # truncate queries if embedding is openai and the query exceeded token limit
+    openai_embedding_limit = 8191
+    if isinstance(embedding_model, OpenAIEmbedding):
+        queries = openai_truncate_by_token(queries, openai_embedding_limit,
+                                           embedding_model.model_name)
+
     # embed query
     embedded_queries = list(map(embedding_model.get_query_embedding, queries))
